Implemented full debian convention
I tried to built a full matching regex according to the debian standard
now see down under link for details. Tough it matches in nearly all
cases it seems that some packages are wrongly named from debian:

- grub2-common_2.02~beta2-22+deb8u1_amd64.deb
- grub-common_2.02~beta2-22+deb8u1_amd64.deb
- grub-pc_2.02~beta2-22+deb8u1_amd64.deb
- grub-pc-bin_2.02~beta2-22+deb8u1_amd64.deb
- debian-archive-keyring_2017.5~deb8u1_all.deb

According to the linked page that should not be a valid debian package
name because they miss the dash or contain a tilde in the debian version
however you like.

https://www.debian.org/doc/manuals/debian-reference/ch02.en.html#_debian_package_file_names